drm/i915: advertise available metrics via sysfs
Each metric set is given a sysfs entry like:

/sys/class/drm/card0/metrics/3d/id
/sys/class/drm/card0/metrics/3d/guid

This allow userspace to enumerate the specific sets that are available
for the current system. The 'id' can be used to open the associated
metric set and the 'guid' can be used as a key to lookup corresponding
meta data, including normalization equations.

Signed-off-by: Robert Bragg <robert@sixbynine.org>